test(kura): keep multipart test launcher without benchmark artifacts
