--- kura/test/e2e/multipart-admission/results.json
@@ -1,191 +1,0 @@
-{
-  "date": "2026-09-09",
-  "platform": "macOS arm64",
-  "baseline_commit": "94d7538febc6f1667479a414362e0aacccd78349",
-  "results": [
-    {
-      "case": "before-128",
-      "capacity": 128,
-      "headroom_mib": 128,
-      "burst": 128,
-      "rounds": 3,
-      "hold_ms": 200,
-      "payload_kib": 64,
-      "start_statuses": {
-        "429": 384
-      },
-      "start_p50_ms": 18.5,
-      "start_p99_ms": 34.38,
-      "completed": 384,
-      "verified_downloads": 0,
-      "ordinary_reads": 45,
-      "ordinary_read_p99_ms": 3.18,
-      "peak_memory_mib": {
-        "kura_jemalloc_allocated_bytes": 3.28,
-        "kura_jemalloc_resident_bytes": 9.78,
-        "kura_memory_transient_reserved_bytes": 0.12,
-        "kura_memory_elastic_transient_reserved_bytes": 0.0
-      },
-      "peak_rss_mib": 158.05,
-      "elapsed_seconds": 4.97
-    },
-    {
-      "case": "after-128",
-      "capacity": 128,
-      "headroom_mib": 128,
-      "burst": 128,
-      "rounds": 3,
-      "hold_ms": 200,
-      "payload_kib": 64,
-      "start_statuses": {
-        "200": 384
-      },
-      "start_p50_ms": 337.42,
-      "start_p99_ms": 685.32,
-      "completed": 768,
-      "verified_downloads": 384,
-      "ordinary_reads": 66,
-      "ordinary_read_p99_ms": 17.85,
-      "peak_memory_mib": {
-        "kura_jemalloc_allocated_bytes": 110.86,
-        "kura_jemalloc_resident_bytes": 258.56,
-        "kura_memory_transient_reserved_bytes": 11.62,
-        "kura_memory_elastic_transient_reserved_bytes": 0.0
-      },
-      "peak_rss_mib": 227.47,
-      "elapsed_seconds": 7.3
-    },
-    {
-      "case": "before-256",
-      "capacity": 256,
-      "headroom_mib": 256,
-      "burst": 256,
-      "rounds": 3,
-      "hold_ms": 200,
-      "payload_kib": 64,
-      "start_statuses": {
-        "429": 768
-      },
-      "start_p50_ms": 41.79,
-      "start_p99_ms": 50.38,
-      "completed": 768,
-      "verified_downloads": 0,
-      "ordinary_reads": 79,
-      "ordinary_read_p99_ms": 4.51,
-      "peak_memory_mib": {
-        "kura_jemalloc_allocated_bytes": 55.92,
-        "kura_jemalloc_resident_bytes": 158.16,
-        "kura_memory_transient_reserved_bytes": 0.12,
-        "kura_memory_elastic_transient_reserved_bytes": 0.0
-      },
-      "peak_rss_mib": 212.89,
-      "elapsed_seconds": 8.57
-    },
-    {
-      "case": "after-256",
-      "capacity": 256,
-      "headroom_mib": 256,
-      "burst": 256,
-      "rounds": 3,
-      "hold_ms": 200,
-      "payload_kib": 64,
-      "start_statuses": {
-        "200": 768
-      },
-      "start_p50_ms": 431.05,
-      "start_p99_ms": 929.09,
-      "completed": 1536,
-      "verified_downloads": 768,
-      "ordinary_reads": 119,
-      "ordinary_read_p99_ms": 12.43,
-      "peak_memory_mib": {
-        "kura_jemalloc_allocated_bytes": 163.13,
-        "kura_jemalloc_resident_bytes": 362.31,
-        "kura_memory_transient_reserved_bytes": 23.69,
-        "kura_memory_elastic_transient_reserved_bytes": 0.0
-      },
-      "peak_rss_mib": 332.33,
-      "elapsed_seconds": 13.26
-    },
-    {
-      "case": "after-1024",
-      "capacity": 1024,
-      "headroom_mib": 1024,
-      "burst": 128,
-      "rounds": 1,
-      "hold_ms": 200,
-      "payload_kib": 64,
-      "start_statuses": {
-        "200": 128
-      },
-      "start_p50_ms": 412.48,
-      "start_p99_ms": 474.69,
-      "completed": 1152,
-      "verified_downloads": 128,
-      "ordinary_reads": 111,
-      "ordinary_read_p99_ms": 7.68,
-      "peak_memory_mib": {
-        "kura_jemalloc_allocated_bytes": 5.02,
-        "kura_jemalloc_resident_bytes": 21.62,
-        "kura_memory_transient_reserved_bytes": 23.88,
-        "kura_memory_elastic_transient_reserved_bytes": 0.0
-      },
-      "peak_rss_mib": 262.06,
-      "elapsed_seconds": 12.21
-    },
-    {
-      "case": "after-overflow",
-      "capacity": 128,
-      "headroom_mib": 128,
-      "burst": 256,
-      "rounds": 1,
-      "hold_ms": 200,
-      "payload_kib": 64,
-      "start_statuses": {
-        "429": 128,
-        "200": 128
-      },
-      "start_p50_ms": 46.67,
-      "start_p99_ms": 652.52,
-      "completed": 256,
-      "verified_downloads": 128,
-      "ordinary_reads": 28,
-      "ordinary_read_p99_ms": 10.04,
-      "peak_memory_mib": {
-        "kura_jemalloc_allocated_bytes": 3.69,
-        "kura_jemalloc_resident_bytes": 10.34,
-        "kura_memory_transient_reserved_bytes": 13.44,
-        "kura_memory_elastic_transient_reserved_bytes": 0.0
-      },
-      "peak_rss_mib": 138.94,
-      "elapsed_seconds": 3.59
-    },
-    {
-      "case": "after-stalled",
-      "capacity": 128,
-      "headroom_mib": 128,
-      "burst": 128,
-      "rounds": 1,
-      "hold_ms": 1500,
-      "payload_kib": 64,
-      "start_statuses": {
-        "429": 128
-      },
-      "start_p50_ms": 1016.84,
-      "start_p99_ms": 1017.73,
-      "completed": 128,
-      "verified_downloads": 0,
-      "ordinary_reads": 28,
-      "ordinary_read_p99_ms": 2.4,
-      "peak_memory_mib": {
-        "kura_jemalloc_allocated_bytes": 3.71,
-        "kura_jemalloc_resident_bytes": 10.55,
-        "kura_memory_transient_reserved_bytes": 5.69,
-        "kura_memory_elastic_transient_reserved_bytes": 0.0
-      },
-      "peak_rss_mib": 97.34,
-      "elapsed_seconds": 3.09
-    }
-  ],
-  "implementation": "FIFO multipart admission with a dedicated pressure-tier signal and cancellation-safe blocking record writes"
-}
